Unify assets.ts
